+++ tests/test_plugins_sqlite_releases.py
@@ -33,8 +33,10 @@ def test_sqlite_releases(reader, requests_mock, data_dir):
     assert feed.link == 'https://www.sqlite.org/changes.html'
 
     (feed_for_update,) = reader._storage.get_feeds_for_update(FeedFilter(FULL_URL))
-    assert feed_for_update.http_etag == 'm6008d7aes58501'
-    assert feed_for_update.http_last_modified == 'Thu, 21 Jan 2021 01:23:58 +0000'
+    assert feed_for_update.caching_info == {
+        'etag': 'm6008d7aes58501',
+        'last-modified': 'Thu, 21 Jan 2021 01:23:58 +0000',
+    }
 
     entries = list(reader.get_entries())
     entry_data = [